# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Athoo Platform API - Website, CRM, Admin Panel
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/waitlist:
    post:
      operationId: submitProviderWaitlist
      tags: [public]
      summary: Submit provider waitlist form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "409":
          description: Already registered
  /public/contact:
    post:
      operationId: submitContactForm
      tags: [public]
      summary: Submit customer contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads with filters
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: source
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: assignedTo
          schema:
            type: integer
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"

    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
        "404":
          description: Not found

    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/notes:
    get:
      operationId: listLeadNotes
      tags: [leads]
      summary: List notes for a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadNote"

    post:
      operationId: createLeadNote
      tags: [leads]
      summary: Add note to lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadNoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNote"
  /leads/bulk-action:
    post:
      operationId: bulkLeadAction
      tags: [leads]
      summary: Bulk action on leads (delete, assign, update status)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkLeadActionInput"
      responses:
        "200":
          description: Action applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"
  /leads/export:
    get:
      operationId: exportLeads
      tags: [leads]
      summary: Export leads as CSV
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: source
          schema:
            type: string
        - in: query
          name: format
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  # ── ADMINS ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, name, email, type, status, source, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        type:
          type: string
          enum: [provider, customer, general]
        status:
          type: string
          enum: [new, contacted, interested, pending, approved, rejected, spam]
        source:
          type: string
          enum: [website, provider_form, customer_form, waitlist, campaign, manual]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        serviceCategory:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadDetail:
      allOf:
        - $ref: "#/components/schemas/Lead"
        - type: object
          properties:
            notes:
              type: array
              items:
                $ref: "#/components/schemas/LeadNote"
            activities:
              type: array
              items:
                $ref: "#/components/schemas/LeadActivity"
    LeadListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    LeadInput:
      type: object
      required: [name, email, type, source]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        city:
          type: string
        type:
          type: string
          enum: [provider, customer, general]
        status:
          type: string
          enum: [new, contacted, interested, pending, approved, rejected, spam]
        source:
          type: string
        serviceCategory:
          type: string
        message:
          type: string
        assignedToId:
          type: integer
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        status:
          type: string
          enum: [new, contacted, interested, pending, approved, rejected, spam]
        assignedToId:
          type: ["integer", "null"]
        serviceCategory:
          type: string
    LeadNote:
      type: object
      required: [id, leadId, content, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        content:
          type: string
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
    LeadNoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    LeadActivity:
      type: object
      required: [id, leadId, action, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        action:
          type: string
        detail:
          type: ["string", "null"]
        adminId:
          type: ["integer", "null"]
        adminName:
          type: ["string", "null"]
        createdAt:
          type: string
    BulkLeadActionInput:
      type: object
      required: [ids, action]
      properties:
        ids:
          type: array
          items:
            type: integer
        action:
          type: string
          enum: [delete, assign, update_status]
        assignedToId:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
    BulkActionResult:
      type: object
      required: [affected]
      properties:
        affected:
          type: integer

    # ── PUBLIC FORMS ──────────────────────────────────────────────────────────────
    WaitlistInput:
      type: object
      required: [name, email, phone, city, serviceCategory]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        city:
          type: string
        serviceCategory:
          type: string
        experience:
          type: string
        message:
          type: string
    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        city:
          type: string
        message:
          type: string

    # ── EMAIL SCHEMAS ────────────────────────────────────────────────────────────
...