# lib/api-zod/src/generated/types/leadActivity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Athoo Platform API - Website, CRM, Admin Panel
 * OpenAPI spec version: 0.1.0
 */

export interface LeadActivity {
  id: number;
  leadId: number;
  action: string;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  adminId?: number | null;
  /** @nullable */
  adminName?: string | null;
  createdAt: string;
}